requirements.txt
babel==2.17.0
backrefs==5.8
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
ghp-import==2.1.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
jinja2==3.1.6
kiwisolver==1.4.8
markdown==3.7
markupsafe==3.0.2
matplotlib==3.9.4
mergedeep==1.3.4
mkdocs-get-deps==0.2.0
mkdocs-material-extensions==1.3.1
mkdocs-material==9.6.9
mkdocs==1.6.1
numpy==2.2.4
packaging==24.2
paginate==0.5.7
pandas==1.5.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pygments==2.19.1
pymdown-extensions==10.14.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml-env-tag==0.1
pyyaml==6.0.2
requests==2.32.3
seaborn==0.12.2
six==1.17.0
urllib3==2.3.0
watchdog==6.0.0
zipp==3.21.0
